Enables warnings -Wconversion in ceedling and fixes most warnings of this

+++ vitis/software/Baremetal/test/uz/uz_matrix/test_uz_matrix_dot_product.c
@@ -13,8 +13,8 @@ void tearDown(void)
 }
 
 void test_uz_matrix_dot_product_two_column_vectors(void){
-    int rows=1;
-    int columns=4;
+    size_t rows=1;
+    size_t columns=4;
     float A_data[4]={1,2,3,4};
     float B_data[4]={1,2,3,4};
     uz_matrix_t* A=uz_matrix_init(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
@@ -24,8 +24,8 @@ void test_uz_matrix_dot_product_two_column_vectors(void){
 }
 
 void test_uz_matrix_dot_product_two_row_vectors(void){
-    int rows=4;
-    int columns=1;
+    size_t rows=4;
+    size_t columns=1;
     float A_data[4]={1,2,3,4};
     float B_data[4]={1,2,3,4};
     uz_matrix_t* A=uz_matrix_init(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
@@ -34,8 +34,8 @@ void test_uz_matrix_dot_product_two_row_vectors(void){
 }
 
 void test_uz_matrix_dot_product_wrong_dimensions(void){
-    int rows=2;
-    int columns=2;
+    size_t rows=2;
+    size_t columns=2;
     float A_data[4]={1,2,3,4};
     float B_data[4]={1,2,3,4};
     uz_matrix_t* A=uz_matrix_init(A_data,UZ_MATRIX_SIZE(A_data),rows, columns );
